# Conflicts:
#	rosen/modules/render_service/core/feature_cfg/feature_param/performance_feature/occlusion_culling_param.cpp
#	rosen/modules/render_service/core/feature_cfg/feature_param/performance_feature/occlusion_culling_param.h
#	rosen/modules/render_service/core/feature_cfg/xml_parser/performance_feature_parser/occlusion_culling_param_parse.cpp
#	rosen/modules/render_service/core/feature_cfg/xml_parser/performance_feature_parser/occlusion_culling_param_parse.h
#	rosen/modules/render_service/core/pipeline/main_thread/rs_uni_render_visitor.cpp
#	rosen/test/render_service/render_service/unittest/feature_cfg/stencil_pixel_occlusion_culling_param_test.cpp

+++ frameworks/text/service/texgine/src/opentype_parser/opentype_basic_type.cpp
@@ -42,7 +42,8 @@ const std::string Tag::Get() const
 int16_t Int16::Get() const
 {
     if (G_ENDIAN.big != '\0') {
-        return ((data & 0x0000ff00) >> 8) | ((data & 0x000000ff) << 8);     // 8 means offset
+        // 8 means offset
+        return ((static_cast<uint16_t>(data) & 0x0000ff00) >> 8) | ((static_cast<uint16_t>(data) & 0x000000ff) << 8);
     } else {
         return data;
     }
